+++ tests/testthat/test_format_fwf.R
@@ -7,18 +7,18 @@ test_that("Export to FWF", {
 })
 
 test_that("Import from FWF (read.fwf)", {
-    expect_true(is.data.frame(import("iris.fwf", widths = c(4,4,4,4,1))))
-    expect_true(is.data.frame(import("iris.fwf", widths = c(4,4,4,4,1), col.names = names(iris))))
-    expect_true(is.data.frame(import("iris.fwf", widths = c(4,4,4,4,1), col.names = names(iris), readr = TRUE)))
-    expect_true(is.data.frame(import("iris.txt", widths = c(4,4,4,4,1), format = "fwf")))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(3,3,3,3,1))))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(3,3,3,3,1), col.names = names(iris))))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(3,3,3,3,1), col.names = names(iris), readr = TRUE)))
+    expect_true(is.data.frame(import("iris.txt", widths = c(3,3,3,3,1), format = "fwf")))
 })
 
 test_that("Import from FWF (read_fwf)", {
-    expect_true(is.data.frame(import("iris.fwf", widths = c(4,4,4,4,1), readr = TRUE)))
-    expect_true(is.data.frame(import("iris.txt", widths = c(4,4,4,4,1), format = "fwf", readr = TRUE)))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(3,3,3,3,1), readr = TRUE)))
+    expect_true(is.data.frame(import("iris.txt", widths = c(3,3,3,3,1), format = "fwf", readr = TRUE)))
     # negative column widths
-    expect_true(is.data.frame(import("iris.fwf", widths = c(-4,4,4,4,1), readr = FALSE)))
-    expect_true(is.data.frame(import("iris.fwf", widths = c(-4,4,4,4,1), readr = TRUE)))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(-3,3,3,3,1), readr = FALSE)))
+    expect_true(is.data.frame(import("iris.fwf", widths = c(-3,3,3,3,1), readr = TRUE)))
 })
 
 test_that("Import from FWF Errors", {
@@ -28,9 +28,8 @@ test_that("Import from FWF Errors", {
     # error on NULL widths
     expect_error(import("iris.fwf", widths = NULL, readr = FALSE))
     # no error on NULL widths w/ readr::read_fwf()
-    suppressWarnings(expect_true(is.data.frame(import("iris.fwf", widths = NULL, readr = TRUE))))
+    expect_true(suppressWarnings(is.data.frame(import("iris.fwf", widths = NULL, readr = TRUE))))
 })
 
 unlink("iris.fwf")
 unlink("iris.txt")
-